requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0
evaluate>=0.4.0
accelerate>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
click>=8.1.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
sentence-transformers>=2.2.0
spacy>=3.6.0
nltk>=3.8.0
psutil>=5.9.0
streamlit>=1.28.0
streamlit-plotly-events>=0.0.6
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6